User account edit functionality added

+++ projects/views.py
@@ -3,6 +3,7 @@ from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
 from .models import Porject
 from .forms import ProjectForm
+from django.contrib import messages
 
 
 def projects(request):
@@ -19,20 +20,30 @@ def project(request, name):
 
 @login_required(login_url='login')
 def createProject(request):
+    profile = request.user.profile
     form = ProjectForm()
 
     if request.method == 'POST':
         form = ProjectForm(request.POST, request.FILES)
         if form.is_valid():
-            form.save()
+            project = form.save(commit=False)
+            project.owner = profile
+            project.save()
             return redirect('projects')
+        else:
+            messages.error(request, 'An error has occurred during registration')
+            print(form.errors)
+
     return render(request, 'projects/project_form.html', {'form': form})
 
 
 @login_required(login_url='login')
 def updateProject(request, pk):
-    project = Porject.objects.get(id=pk)
-
+    profile = request.user.profile
+    project = profile.project_set.get(id=pk)
+    if project is None:
+        messages.error(request, "Project not found!")
+        return redirect('projects')
     form = ProjectForm(instance=project)
 
     if request.method == 'POST':
@@ -45,8 +56,13 @@ def updateProject(request, pk):
 
 @login_required(login_url='login')
 def deleteProject(request, pk):
-    project = Porject.objects.get(id=pk)
+    profile = request.user.profile
+
+    project = profile.porject_set.get(id=pk)
+    if project is None:
+        messages.error(request, 'Project not found')
+        return redirect('projects')
     if request.method == 'POST':
         project.delete();
         return redirect('projects')
-    return render(request, 'projects/delete_template.html', {'object': project})
+    return render(request, 'delete_template.html', {'object': project})
